backend/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
flask-migrate
